+++ packages/services/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sharelyai/services
 
+## 0.0.28
+
+### Patch Changes
+
+- update chat references
+
 ## 0.0.27
 
 ### Patch Changes