+++ packages/ui-browse/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @sharelyai/ui-browse
 
+## 0.0.28
+
+### Patch Changes
+
+- update chat references
+- Updated dependencies
+  - @sharelyai/ui-search@0.0.28
+  - @sharelyai/ui-shared@0.0.28
+  - @sharelyai/services@0.0.28
+
 ## 0.0.27
 
 ### Patch Changes